requirements.txt
Django==1.11.2
Pillow
django-crispy-forms==1.6.1
python-dateutil
django-registration-redux==1.7
MySQL-python
python-social-auth==0.2.14